requirements.txt
Flask==3.0.3
openpyxl==3.1.5
db-sqlite3==0.0.1
pillow==10.4.0
gunicorn==23.0.0
xlsxwriter==3.2.0